requirements.txt
# Core HTTP and async support
httpx>=0.24.0
aiohttp>=3.8.0
requests>=2.28.0

# Provider-specific libraries
openai>=1.0.0
anthropic>=0.7.0
ollama>=0.1.0

# Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Utilities
typing-extensions>=4.5.0
tenacity>=8.2.0  # For retry logic

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0